+++ README.md
@@ -1,8 +1,8 @@
-# eslint-plugin-detect-unwanted-words
+# eslint-plugin-detect-bad-words
 
-![Node.js CI](https://github.com/darwintantuco/eslint-plugin-detect-unwanted-words/workflows/Node.js%20CI/badge.svg?branch=master)
+![Node.js CI](https://github.com/darwintantuco/eslint-plugin-detect-bad-words/workflows/Node.js%20CI/badge.svg?branch=master)
 
-Detect unwanted words in code and comments
+Detect bad words in code and comments
 
 ![](demo.png)
 
@@ -10,36 +10,36 @@ Detect unwanted words in code and comments
 
 You'll first need to install [ESLint](http://eslint.org):
 
-Next, install `eslint-plugin-detect-unwanted-words`:
+Next, install `eslint-plugin-detect-bad-words`:
 
 ### npm
 
 ```
-$ npm install eslint-plugin-detect-unwanted-words --save-dev
+$ npm install eslint-plugin-detect-bad-words --save-dev
 ```
 
 ### yarn
 
 ```
-$ yarn add eslint-plugin-detect-unwanted-words --dev
+$ yarn add eslint-plugin-detect-bad-words --dev
 ```
 
 ## Usage
 
-Add `detect-unwanted-words` to the plugins section of your `.eslintrc` configuration file. You can omit the `eslint-plugin-` prefix:
+Add `detect-bad-words` to the plugins section of your `.eslintrc` configuration file. You can omit the `eslint-plugin-` prefix:
 
 ```json
 {
-  "plugins": ["detect-unwanted-words"]
+  "plugins": ["detect-bad-words"]
 }
 ```
 
-Define unwanted words under settings section.
+Define custom bad words under settings section.
 
 ```json
 {
   "settings": {
-    "unwantedWords": ["fck", "sht", "tite", "trust me"]
+    "customBadWords": ["fck", "sht", "tite", "trust me"]
   }
 }
 ```
@@ -49,8 +49,8 @@ Then configure the rules you want to use under the rules section.
 ```json
 {
   "rules": {
-    "detect-unwanted-words/detect-unwanted-words-in-code": "error",
-    "detect-unwanted-words/detect-unwanted-words-in-comments": "error"
+    "detect-bad-words/in-code": "error",
+    "detect-bad-words/in-comments": "error"
   }
 }
 ```
