+++ lib/index.js
@@ -1,5 +1,5 @@
 /**
- * @fileoverview Detect unwanted words
+ * @fileoverview Detect bad words
  * @author Darwin Tantuco
  */
 'use strict'
