+++ lib/rules/in-code.js
@@ -1,17 +1,17 @@
 'use strict'
 
-const { searchForUnwantedWords, isEmpty } = require('../util')
+const { searchForCustomBadWords, isEmpty } = require('../util')
 
 module.exports = {
   create: (context) => {
     const settings = context.settings || {}
-    const unwantedWords = settings.unwantedWords || []
+    const customBadWords = settings.customBadWords || []
 
     return {
       Literal(node) {
         if (typeof node.value !== 'string') return
 
-        const result = searchForUnwantedWords(unwantedWords, node.value)
+        const result = searchForCustomBadWords(customBadWords, node.value)
 
         if (!isEmpty(result)) {
           context.report({
@@ -26,7 +26,7 @@ module.exports = {
       JSXText(node) {
         if (typeof node.value !== 'string') return
 
-        const result = searchForUnwantedWords(unwantedWords, node.value)
+        const result = searchForCustomBadWords(customBadWords, node.value)
 
         if (!isEmpty(result)) {
           context.report({
