+++ lib/rules/in-comments.js
@@ -1,15 +1,15 @@
 'use strict'
 
-const { searchForUnwantedWords, isEmpty } = require('../util')
+const { searchForCustomBadWords, isEmpty } = require('../util')
 
 module.exports = {
   create: (context) => {
     const settings = context.settings || {}
-    const unwantedWords = settings.unwantedWords || []
+    const customBadWords = settings.customBadWords || []
     const sourceCode = context.getSourceCode()
 
     const validate = (comment) => {
-      const result = searchForUnwantedWords(unwantedWords, comment.value)
+      const result = searchForCustomBadWords(customBadWords, comment.value)
       if (!isEmpty(result)) {
         context.report({
           node: null,
