+++ lib/util.js
@@ -2,8 +2,8 @@
 
 const badwordsObject = require('badwords/object')
 
-const searchForUnwantedWords = (unwantedWords, value) => {
-  return unwantedWords.reduce((acc, word) => {
+const searchForCustomBadWords = (customBadWords, value) => {
+  return customBadWords.reduce((acc, word) => {
     const pattern = new RegExp('\\b' + escapeString(word) + '\\b', 'i')
 
     if (pattern.test(value)) {
@@ -24,6 +24,6 @@ const escapeString = (string) =>
   string.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&')
 
 module.exports = {
-  searchForUnwantedWords,
+  searchForCustomBadWords,
   isEmpty,
 }
